self bootstrapping up and running
I can't tell whether all the extra elisp that is needed is worth it,
it seems like there should be an easier way to ensure that some elisp
functions have been defined before running certain blocks, maybe the
modline is a simpler way to do this? On the other hand this is only
one step removed from asking someone to curl pipe sh something, and
apparently having an independent checksum that isn't in the file
itself is still too much to ask for many users. Ah well.

Also regularized the directory structure for tangled files to match
what I do in other projects.