a bit of thinking on how to orgstrap system packages
Package managers are non-trivial to deal with, whether they are system
or Emacs. Section per manager almost makes sense because there are
many quriks that have to be sorted out. The diversity of ways to get
from the name of a package to the ability to use its functionality
makes it quite difficult for a single piece of software to be self
documenting without basically having an entire section with nothing
but special cases. At the other end of the spectrum, if everyone has
some compatibility layer, standarized container, vm, etc. then a
single working example could be provided. It would still be nice
to be able to have all the different steps required to install and
run something correctly in a given environment documented in the
file. As it stands, source code is entirely at the whims of the
distro maintainers simply because there are no standard conventions
for stuff like this.